Move comment attach/print code
