+++ src/language-js/print/block.js
@@ -1,4 +1,4 @@
-import { printDanglingComments } from "../../main/comments.js";
+import { printDanglingComments } from "../../main/comments/print.js";
 import { isNonEmptyArray } from "../../common/util.js";
 import { hardline, indent } from "../../document/builders.js";
 import {
