+++ src/language-js/print/member-chain.js
@@ -1,4 +1,4 @@
-import { printComments } from "../../main/comments.js";
+import { printComments } from "../../main/comments/print.js";
 import {
   isNextLineEmpty as isNextLineEmptyAfterIndex,
   getNextNonSpaceNonCommentCharacterIndex,
